clarifications in readme
Copying and pasting from the jade docs means that if the jade docs
change, we'd need to notice and update. I changed it to link directly to
the jade docs, where the options are clearly described, just listing
specific differences with this plugin.

I also added an example of `locals` working to the example below, since
that's the only option that is different from the docs.